backend/requirements.txt
flask
flask-cors
pillow
werkzeug
gunicorn 